backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12
pydantic-settings==2.5.2
httpx==0.27.2
pymongo==4.10.1
email-validator==2.2.0
